// src/app/model/Employee.ts
export class Employee {


    employeeId: number = 0;
    employeeFirstName: string = '';
    employeeLastName: string = '';
    email: string = '';
    phone: number = 0;
    salary: number = 0;
    username: string = '';
    password: string = '';
    status: boolean = true;
    userId: number = 0;
}